# ENH: sparse: efficient arithmetic operations for DIA format