--- a/scipy/sparse/_generate_sparsetools.py
+++ b/scipy/sparse/_generate_sparsetools.py
@@ -112,6 +112,7 @@ coo_matvec_nd       v llIITT*T
 coo_matmat_dense    v llIITT*T
 coo_matmat_dense_nd v lllIIITT*T
 dia_matvec          v iiiiITT*T
+dia_matvecs         v iiiiITiT*T
 cs_graph_components i iII*I
 """