--- a/scipy/sparse/sparsetools/dia.h
+++ b/scipy/sparse/sparsetools/dia.h
@@ -2,6 +2,120 @@
 #define __DIA_H__
 
 #include <algorithm>
+#include <vector>
+
+using std::min, std::max;
+
+template <class I, class T>
+T min_el(const T vec[], I len) {
+  return *std::min_element(vec, vec + len);
+}
+
+template <class I, class T>
+T max_el(const T vec[], I len) {
+  return *std::max_element(vec, vec + len);
+}
+
+
+/*
+ * Compute DIA matrix output = A * B for DIA matrices A, B
+ *
+ *
+ * Input Arguments:
+ *   I  A_rows               - number of rows in A
+ *   I  A_cols               - number of columns in A
+ *   I  A_diags              - number of diagonals in A
+ *   I  A_L                  - length of each diagonal in A
+ *   I  A_offsets[A_diags]   - diagonal offsets in A
+ *   T  A_data[A_diags,A_L]  - diagonals data of A
+ *   I  B_cols               - number of columns in B
+ *   I  B_diags              - number of diagonals in B
+ *   I  B_L                  - length of each diagonal in B
+ *   I  B_offsets[B_diags]   - diagonal offsets in B
+ *   T  B_data[B_diags,B_L]  - diagonals data of B
+ *
+ * Output Arguments:
+ *   V  offsets              - diagonal offsets
+ *   W  data                 - diagonals data
+ *
+ * Note:
+ *   Number of diagonals in output (diags) is the length of output offsets;
+ *   length of each diagonal (L) equals B_L; data dimensions are [diags,L]
+ *   Negative offsets correspond to lower diagonals
+ *   Positive offsets correspond to upper diagonals
+ *
+ */
+template <class I, class T>
+void dia_matmat(const I A_rows,
+                const I A_cols,
+                const I A_diags,
+                const I A_L,
+                const I A_offsets[],
+                const T A_data[],
+                const I B_cols,
+                const I B_diags,
+                const I B_L,
+                const I B_offsets[],
+                const T B_data[],
+                std::vector<I>* offsets,
+                std::vector<T>* data)
+{
+    const I B_rows = A_cols,
+            rows = A_rows, cols = B_cols,
+            L = min(cols, B_L);
+
+    // range of combinations of input offsets
+    const I min_map_ofs = min_el(A_offsets, A_diags) + min_el(B_offsets, B_diags),
+            max_map_ofs = max_el(A_offsets, A_diags) + max_el(B_offsets, B_diags);
+    // limits for output offsets
+    const I min_ofs = max(min_map_ofs, 1 - rows),
+            max_ofs = min(max_map_ofs, L - 1);
+    // mark output offsets
+    std::vector<I> buf(max_map_ofs - min_map_ofs + 1);
+    // (alias to buf indexable from [min_map_ofs] to [max_map_ofs];
+    //  only [min_ofs] to [max_ofs] will be used later)
+    I* ofs_map = buf.data() - min_map_ofs;
+    for (I i = 0; i < A_diags; ++i)
+        for (I j = 0; j < B_diags; ++j)
+            ofs_map[A_offsets[i] + B_offsets[j]] = I(true);
+    // enumerate marks
+    offsets->resize(max_ofs - min_ofs + 1);
+    I N_ofs = 0;
+    for (I ofs = min_ofs; ofs <= max_ofs; ++ofs)
+        if (ofs_map[ofs]) {
+            (*offsets)[N_ofs] = ofs;
+            ofs_map[ofs] = N_ofs;
+            ++N_ofs;
+        }
+    offsets->resize(N_ofs);
+
+    // allocate output diagonals, filled with zeros
+    data->resize(N_ofs * L);
+    // loop over diagonals in B
+    for (I B_i = 0; B_i < B_diags; ++B_i) {
+        const I B_ofs = B_offsets[B_i];
+        const T* B_diag_r = B_data + npy_intp(B_L) * B_i + B_ofs; // row-indexed
+        // span of data within current (row-indexed) B diagonal
+        const I B_j_beg = max(-B_ofs, I(0)),
+                B_j_end = min(min(B_cols, B_L) - B_ofs, B_rows);
+        // loop over diagonals in A
+        for (I A_i = 0; A_i < A_diags; ++A_i) {
+            const I A_ofs = A_offsets[A_i];
+            const T* A_diag_c = A_data + npy_intp(A_L) * A_i; // column-indexed
+            // output offset and corresponding diagonal
+            const I ofs = A_ofs + B_ofs;
+            if (ofs < min_ofs or ofs > max_ofs)
+                continue;
+            T* diag_r = data->data() + npy_intp(L) * ofs_map[ofs] + B_ofs; // row-indexed
+            // overlapping span of data within current B and A diagonals
+            const I j_beg = max(B_j_beg, A_ofs),
+                    j_end = min({B_j_end, min(A_cols, A_L), A_rows + A_ofs});
+            // add partial product to output
+            for (I j = j_beg; j < j_end; ++j)
+                diag_r[j] += A_diag_c[j] * B_diag_r[j];
+        }
+    }
+}
 
 
 /*