#### Reference issue
Partially addresses gh-21039.

#### What does this implement/fix?
Implements most important arithmetic operations with sparse arrays in DIA format. Previously most operations used fall-back conversion to CSR format with very poor overall efficiency.

For comparison, the standard benchmarks (``python dev.py bench -s sparse.Arithmetic``):

##### Before
```
sparse.Arithmetic.time_arithmetic                        
======== ==== ========== ========== ========== ==========
--                                 op                    
------------- -------------------------------------------
 format   XY   __add__    __sub__    multiply   __mul__  
======== ==== ========== ========== ========== ==========
  csr     AA   1.95±0ms   1.15±0ms   1.97±0ms   10.4±0ms 
  csr     AB   7.12±0ms   13.8±0ms   7.64±0ms   21.1±0ms 
  csr     BA   6.89±0ms   7.05±0ms   6.52±0ms   24.4±0ms 
  csr     BB   7.06±0ms   5.21±0ms   6.98±0ms   46.0±0ms 
  csc     AA   3.63±0ms   6.53±0ms   3.49±0ms   10.2±0ms 
  csc     AB   6.55±0ms   6.45±0ms   7.40±0ms   24.4±0ms 
  csc     BA   6.40±0ms   6.43±0ms   7.31±0ms   20.8±0ms 
  csc     BB   7.04±0ms   5.65±0ms   7.07±0ms   44.9±0ms 
  coo     AA   18.5±0ms   15.4±0ms   17.9±0ms   27.2±0ms 
  coo     AB   25.9±0ms   26.0±0ms   26.2±0ms   42.0±0ms 
  coo     BA   25.8±0ms   29.4±0ms   26.5±0ms   49.8±0ms 
  coo     BB   35.0±0ms   30.8±0ms   33.0±0ms   72.9±0ms 
  dia     AA   1.44±0ms   28.2±0ms   29.3±0ms   38.2±0ms 
  dia     AB   2.37±0ms   46.7±0ms   46.8±0ms   61.5±0ms 
  dia     BA   2.16±0ms   46.0±0ms   42.3±0ms   66.9±0ms 
  dia     BB   1.51±0ms   64.6±0ms   66.8±0ms   108±0ms  
======== ==== ========== ========== ========== ==========
```
(Only addition was implemented.)

##### After
```
sparse.Arithmetic.time_arithmetic                        
======== ==== ========== ========== ========== ==========
--                                 op                    
------------- -------------------------------------------
 format   XY   __add__    __sub__    multiply   __mul__  
======== ==== ========== ========== ========== ==========
  csr     AA   1.90±0ms   1.19±0ms   1.94±0ms   10.1±0ms 
  csr     AB   6.86±0ms   6.85±0ms   6.39±0ms   20.2±0ms 
  csr     BA   6.63±0ms   6.60±0ms   6.14±0ms   23.9±0ms 
  csr     BB   6.87±0ms   5.16±0ms   7.22±0ms   44.8±0ms 
  csc     AA   3.46±0ms   2.85±0ms   3.60±0ms   10.3±0ms 
  csc     AB   6.72±0ms   6.51±0ms   6.79±0ms   24.3±0ms 
  csc     BA   6.54±0ms   6.70±0ms   6.69±0ms   20.4±0ms 
  csc     BB   7.01±0ms   5.65±0ms   7.02±0ms   42.5±0ms 
  coo     AA   18.1±0ms   15.4±0ms   17.7±0ms   26.0±0ms 
  coo     AB   26.7±0ms   26.8±0ms   27.4±0ms   42.9±0ms 
  coo     BA   27.0±0ms   27.0±0ms   27.6±0ms   46.4±0ms 
  coo     BB   35.7±0ms   32.6±0ms   35.3±0ms   73.7±0ms 
  dia     AA   596±0μs    596±0μs    1.42±0ms   7.07±0ms 
  dia     AB   2.38±0ms   3.17±0ms   1.41±0ms   13.0±0ms 
  dia     BA   2.42±0ms   2.44±0ms   1.49±0ms   12.8±0ms 
  dia     BB   1.26±0ms   1.45±0ms   5.74±0ms   26.0±0ms 
======== ==== ========== ========== ========== ==========
```
(Now DIA is the most efficient in all respects, as expected.)

DIA @ dense is also implemented and has efficiency similar to CSR @ dense, so the overall performance is improved by avoiding the DIA → COO → CSR conversion.

#### Additional information
<!--Any additional information you think is important.-->
I'm not familiar with SciPy internals and didn't find technical notes about `scipy.sparse` in particular (please point out if they exist), so there are some details that might need attention:
* Are `dtype` promotions done correctly?
* Should indexing `dtype` be specified? (The existing code sometimes uses defaults and sometimes `intc`, even though `intp` should be more time-efficient under Windows, see [numpy 15513](https://github.com/numpy/numpy/issues/15513#event-13468943685).)
* Is `.ravel()` needed for arrays passed to wrapped C++ functions? (Again, seems to work without it, but some existing code uses it.)
* Should error messages be more informative and consistent? I mean, there's been some work in gh-21018, but to a very limited extent.

To complete gh-21039, divisions and comparisons also need to be implemented, but in order to do them, I need to know 
the conventions (what exactly should happen with operations on implicit zeros?).