@perimosocordiae 2024-07-24
Thanks for the PR! The improvements for subtraction here make me think we might want to apply a similar technique more broadly across the various formats. If you want to split that part out into a separate PR, we can merge that part more quickly.

I haven't read through the matmul changes closely yet, but overall the benchmarks and tests look very nice.