@perimosocordiae 2024-07-24
Also FYI: `dev.py bench` supports a comparison mode with the `--compare main` flag. This can make it easier to see where the big changes occur.