@MikhailRyazanov 2024-07-24
Thanks, @perimosocordiae! Please continue the review whenever you have time.

I suspect that the addition code that I've generalized here to subtraction is also potentially generalizable to some comparisons (and or/xor, if these to be implemented), on which I was planning to work after this gets reviewed. It would then make more sense to port the complete solution to other formats...

As for the benchmarks comparison, my goal was not so much to speed-up DIA on its own but to make it faster than CSR (as it should be theoretically), hence the way I've presented the results.