@MikhailRyazanov 2025-05-15
Thanks, @tylerjereddy!

> The rebased version of this feature branch still passes the full testsuite locally.

The last time I've checked the status of this PR, the only “conflict” was that another PR added an unrelated signature in `_generate_sparsetools.py`, so the rebase looked trivial and I wasn't sure whether I need to do it myself. Let me know if any other updates/changes are needed.